requirements.txt
absl-py==2.3.1
annotated-types==0.7.0
asttokens @ file:///home/conda/feedstock_root/build_artifacts/asttokens_1763409923949/work
attrs==25.4.0
blinker==1.9.0
blis==1.3.3
catalogue==2.0.10
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
cloudpathlib==0.23.0
colorama @ file:///home/conda/feedstock_root/build_artifacts/colorama_1733218098505/work
comm @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_comm_1753453984/work
confection==0.1.5
contourpy==1.3.2
cryptography==46.0.3
cycler==0.12.1
cymem==2.0.13
debugpy @ file:///D:/bld/bld/rattler-build_debugpy_1758162070/work
decorator @ file:///home/conda/feedstock_root/build_artifacts/decorator_1740384970518/work
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl#sha256=1932429db727d4bff3deed6b34cfc05df17794f4a52eeb26cf8928f7c1a0fb85
exceptiongroup @ file:///home/conda/feedstock_root/build_artifacts/exceptiongroup_1746947292760/work
executing @ file:///home/conda/feedstock_root/build_artifacts/executing_1756729339227/work
filelock==3.20.0
Flask==3.1.2
flatbuffers==25.9.23
fonttools==4.60.1
fsspec==2025.10.0
gensim==4.4.0
idna==3.11
importlib_metadata @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_importlib-metadata_1747934053/work
ipykernel @ file:///D:/bld/bld/rattler-build_ipykernel_1761567966/work
ipython @ file:///D:/bld/bld/rattler-build_ipython_1748711254/work
itsdangerous==2.2.0
jax==0.6.2
jaxlib==0.6.2
jedi @ file:///home/conda/feedstock_root/build_artifacts/jedi_1733300866624/work
Jinja2==3.1.6
joblib==1.5.2
jupyter_client @ file:///home/conda/feedstock_root/build_artifacts/jupyter_client_1733440914442/work
jupyter_core @ file:///D:/bld/bld/rattler-build_jupyter_core_1760643889/work
kiwisolver==1.4.9
MarkupSafe==3.0.3
matplotlib==3.10.7
matplotlib-inline @ file:///home/conda/feedstock_root/build_artifacts/matplotlib-inline_1761214490209/work
mediapipe==0.10.21
ml_dtypes==0.5.4
mpmath==1.3.0
murmurhash==1.0.15
nest_asyncio @ file:///home/conda/feedstock_root/build_artifacts/nest-asyncio_1733325553580/work
networkx==3.4.2
nltk==3.9.2
numpy==1.26.4
opencv-contrib-python==4.11.0.86
opencv-python==4.12.0.88
opt_einsum==3.4.0
packaging @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_packaging_1745345660/work
pandas==2.3.3
parso @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_parso_1755974222/work
pdfminer.six==20251107
pdfplumber==0.11.8
pickleshare @ file:///home/conda/feedstock_root/build_artifacts/pickleshare_1733327343728/work
pillow==12.0.0
platformdirs @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_platformdirs_1759953252/work
polars==1.35.2
polars-runtime-32==1.35.2
preshed==3.0.12
prompt_toolkit @ file:///home/conda/feedstock_root/build_artifacts/prompt-toolkit_1756321756983/work
protobuf==4.25.8
psutil @ file:///D:/bld/bld/rattler-build_psutil_1762092906/work
pure_eval @ file:///home/conda/feedstock_root/build_artifacts/pure_eval_1733569405015/work
pycparser==2.23
pydantic==2.12.4
pydantic_core==2.41.5
Pygments @ file:///home/conda/feedstock_root/build_artifacts/pygments_1750615794071/work
pyparsing==3.2.5
PyPDF2==3.0.1
pypdfium2==5.1.0
python-dateutil @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_python-dateutil_1751104122/work
pytz==2025.2
pywin32 @ file:///D:/bld/bld/rattler-build_pywin32_1756487147/work
PyYAML==6.0.3
pyzmq @ file:///D:/bld/bld/rattler-build_pyzmq_1757387021/work
regex==2025.11.3
requests==2.32.5
scikit-learn==1.7.2
scipy==1.15.3
seaborn==0.13.2
sentencepiece==0.2.1
six @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_six_1753199211/work
smart_open==7.5.0
sounddevice==0.5.3
spacy==3.8.11
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.5.2
stack_data @ file:///home/conda/feedstock_root/build_artifacts/stack_data_1733569443808/work
sympy==1.14.0
thinc==8.3.10
threadpoolctl==3.6.0
torch==2.9.1
torchvision==0.24.1
tornado @ file:///D:/bld/tornado_1762506734045/work
tqdm==4.67.1
traitlets @ file:///home/conda/feedstock_root/build_artifacts/traitlets_1733367359838/work
typer-slim==0.20.0
typing-inspection==0.4.2
typing_extensions @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_typing_extensions_1756220668/work
tzdata==2025.2
ultralytics==8.3.229
ultralytics-thop==2.0.18
urllib3==2.5.0
wasabi==1.1.3
wcwidth @ file:///home/conda/feedstock_root/build_artifacts/wcwidth_1758622279606/work
weasel==0.4.3
Werkzeug==3.1.3
wrapt==2.0.1
zipp @ file:///home/conda/feedstock_root/build_artifacts/zipp_1749421620841/work
